refactor: move buildDescriminatedAuth to KintoneRESTApiClient
