+++ packages/rest-api-client/src/platform/index.ts
@@ -1,8 +1,10 @@
+import { DiscriminatedAuth } from "./../KintoneRestAPIClient";
 type PlatformDeps = {
   readFileFromPath: (
     filePath: string
   ) => Promise<{ name: string; data: unknown }>;
   getRequestToken: () => string;
+  getDefaultAuth: () => DiscriminatedAuth;
   buildPlatformDependentConfig: (params: object) => object;
   buildHeaders: () => Record<string, string>;
 };
@@ -12,6 +14,9 @@ export const platformDeps: PlatformDeps = {
     throw new Error("not implemented");
   },
   getRequestToken: () => {
+    throw new Error("not implemented");
+  },
+  getDefaultAuth: () => {
     throw new Error("not implemented");
   },
   buildPlatformDependentConfig: () => {
@@ -25,6 +30,7 @@ export const platformDeps: PlatformDeps = {
 export const injectPlatformDeps = (deps: Partial<PlatformDeps>) => {
   platformDeps.readFileFromPath = deps.readFileFromPath!;
   platformDeps.getRequestToken = deps.getRequestToken!;
+  platformDeps.getDefaultAuth = deps.getDefaultAuth!;
   platformDeps.buildPlatformDependentConfig = deps.buildPlatformDependentConfig!;
   platformDeps.buildHeaders = deps.buildHeaders!;
 };
